+++ Yeti/Controllers/Feeds/NewFolderVC.h
@@ -6,12 +6,12 @@
 //  Copyright © 2018 Dezine Zync Studios. All rights reserved.
 //
 
-#import "NewFeedVC.h"
+//#import "NewFeedVC.h"
 
 #import "Folder.h"
 #import "SidebarVC.h"
 
-@interface NewFolderVC : NewFeedVC
+@interface NewFolderVC : UIViewController
 
 + (UINavigationController * _Nonnull)instanceWithFolder:(Folder * _Nonnull)folder indexPath:(NSIndexPath * _Nonnull)indexPath;